+++ presto-main/src/main/java/com/facebook/presto/operator/LookupSource.java
@@ -27,7 +27,7 @@ public interface LookupSource
 
     int getJoinPositionCount();
 
-    long getJoinPosition(int position, Page page, int rawHash);
+    long getJoinPosition(int position, Page page, long rawHash);
 
     long getJoinPosition(int position, Page page);
