+++ presto-main/src/main/java/com/facebook/presto/type/JsonType.java
@@ -51,7 +51,7 @@ public class JsonType
     }
 
     @Override
-    public int hash(Block block, int position)
+    public long hash(Block block, int position)
     {
         return block.hash(position, 0, block.getLength(position));
     }
