+++ presto-spi/src/test/java/com/facebook/presto/spi/type/TestingIdType.java
@@ -56,7 +56,7 @@ public class TestingIdType
     }
 
     @Override
-    public int hash(Block block, int position)
+    public long hash(Block block, int position)
     {
         long value = block.getLong(position, 0);
         return (int) (value ^ (value >>> 32));
